refactor: refactor with authorization, change current account type
